organize input components
